Cleanup SendMode
